+++ include/gammu-memory.h
@@ -1,7 +1,7 @@
 /**
  * \file gammu-memory.h
  * \author Michal Čihař
- * 
+ *
  * Memory data and functions.
  */
 #ifndef __gammu_memory_h
@@ -19,7 +19,7 @@
  * Enum defines ID for various phone and SIM memories.
  * Phone modules can translate them to values specific for concrete models.
  * Two letter codes (excluding VM and SL) are from GSM 07.07.
- * 
+ *
  * \ingroup Memory
  */
 typedef enum {
@@ -70,7 +70,7 @@ typedef enum {
 /**
  * Structure contains info about number of used/free entries in phonebook
  * memory.
- * 
+ *
  * \ingroup Memory
  */
 typedef struct {
@@ -91,7 +91,7 @@ typedef struct {
 /**
  * Type of specific phonebook entry. In parenthesis is specified in which
  * member of @ref GSM_SubMemoryEntry value is stored.
- * 
+ *
  * \ingroup Memory
  */
 typedef enum {
@@ -228,7 +228,7 @@ typedef enum {
 	 */
 	PBK_Text_UserID,
 	/**
-	 * Length of call (Number) 
+	 * Length of call (Number)
 	 */
 	PBK_CallLength,
 	/**
@@ -236,15 +236,15 @@ typedef enum {
 	 */
 	PBK_Text_LUID,
 	/**
-	 * Date of last modification 
+	 * Date of last modification
 	 */
 	PBK_LastModified,
 	/**
-	 * Nick name (Text) 
+	 * Nick name (Text)
 	 */
 	PBK_Text_NickName,
 	/**
-	 * Formal name (Text) 
+	 * Formal name (Text)
 	 */
 	PBK_Text_FormalName,
 	/**
@@ -283,7 +283,7 @@ typedef enum {
 
 /**
  * One value of phonebook memory entry.
- * 
+ *
  * \ingroup Memory
  */
 typedef struct {
@@ -318,7 +318,7 @@ typedef struct {
 
 /**
  * Structure for saving phonebook entries.
- * 
+ *
  * \ingroup Memory
  */
 typedef struct {
@@ -342,7 +342,7 @@ typedef struct {
 
 /**
  * Structure for saving speed dials.
- * 
+ *
  * \ingroup Memory
  */
 typedef struct {
@@ -367,7 +367,7 @@ typedef struct {
 /**
  * Gets memory (phonebooks or calls) status (eg. number of used and
  * free entries).
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -379,7 +379,7 @@ GSM_Error GSM_GetMemoryStatus(GSM_StateMachine * s, GSM_MemoryStatus * status);
 /**
  * Reads entry from memory (phonebooks or calls). Which entry should
  * be read is defined in entry.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -391,7 +391,7 @@ GSM_Error GSM_GetMemory(GSM_StateMachine * s, GSM_MemoryEntry * entry);
 /**
  * Reads entry from memory (phonebooks or calls). Which entry should
  * be read is defined in entry. This can be easily used for reading all entries.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -403,7 +403,7 @@ GSM_Error GSM_GetNextMemory(GSM_StateMachine * s, GSM_MemoryEntry * entry,
 
 /**
  * Sets memory (phonebooks or calls) entry.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -414,7 +414,7 @@ GSM_Error GSM_SetMemory(GSM_StateMachine * s, GSM_MemoryEntry * entry);
 
 /**
  * Deletes memory (phonebooks or calls) entry.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -425,7 +425,7 @@ GSM_Error GSM_AddMemory(GSM_StateMachine * s, GSM_MemoryEntry * entry);
 
 /**
  * Deletes memory (phonebooks or calls) entry.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -436,7 +436,7 @@ GSM_Error GSM_DeleteMemory(GSM_StateMachine * s, GSM_MemoryEntry * entry);
 
 /**
  * Deletes all memory (phonebooks or calls) entries of specified type.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -447,7 +447,7 @@ GSM_Error GSM_DeleteAllMemory(GSM_StateMachine * s, GSM_MemoryType MemoryType);
 
 /**
  * Gets speed dial.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -458,7 +458,7 @@ GSM_Error GSM_GetSpeedDial(GSM_StateMachine * s, GSM_SpeedDial * Speed);
 
 /**
  * Sets speed dial.
- * 
+ *
  * \param s State machine pointer.
  *
  * \return Error code.
@@ -470,7 +470,7 @@ GSM_Error GSM_SetSpeedDial(GSM_StateMachine * s, GSM_SpeedDial * Speed);
 /**
  * Returns name of entry. It might be possibly concatenated from first
  * and last names.
- * 
+ *
  * \param entry Entry to process.
  *
  * \return Static unicode string containing name.
@@ -481,7 +481,7 @@ unsigned char *GSM_PhonebookGetEntryName(GSM_MemoryEntry * entry);
 
 /**
  * Finds default name, number and group for entry.
- * 
+ *
  * \param entry Entry to process.
  * \param Name Output index of name.
  * \param Number Output index of number.
@@ -522,7 +522,7 @@ typedef enum {
  *
  * \ingroup Memory
  */
-void GSM_EncodeVCARD(char *Buffer, int *Length, GSM_MemoryEntry * pbk,
+void GSM_EncodeVCARD(char *Buffer, size_t *Length, GSM_MemoryEntry * pbk,
 		     bool header, GSM_VCardVersion Version);
 
 /**
